+++ src/Proxy/Observer.cs
@@ -1,0 +1,7 @@
+﻿namespace Proxy
+{
+    public class Observer
+    {
+        public string Subject { get; set; }
+    }
+}